Various Bugfixes and improvements
- Now the VPI handles are stored internally in SimVpi to avoid getting invalid values in re-run simulations
- Thread locking on VpiBackend creation and .instanciate
- VpiBackend.instanciate waits the simulation to be ready before returning
- Elaboration and run step was splitted in GhdlBackend. Now elaboration is executed in  .analyze

+++ sim/src/main/scala/spinal/sim/SimVerilator.scala
@@ -6,23 +6,23 @@ object SimVerilator{
 
 class SimVerilator(backend : VerilatorBackend, handle : Long) extends SimRaw(){
   override def getInt(signal : Signal) : Int = {
-    assert(signal.id.toInt != -1, "You can't access this signal in the simulation, as it isn't public")
-    signal.dataType.raw64ToInt(backend.nativeInstance.getU64(handle, signal.id.toInt), signal : Signal)
+    assert(signal.id != -1, "You can't access this signal in the simulation, as it isn't public")
+    signal.dataType.raw64ToInt(backend.nativeInstance.getU64(handle, signal.id), signal : Signal)
   }
   override def getLong(signal : Signal) : Long = {
-    assert(signal.id.toInt != -1, "You can't access this signal in the simulation, as it isn't public")
-    signal.dataType.raw64ToLong(backend.nativeInstance.getU64(handle, signal.id.toInt), signal : Signal)
+    assert(signal.id != -1, "You can't access this signal in the simulation, as it isn't public")
+    signal.dataType.raw64ToLong(backend.nativeInstance.getU64(handle, signal.id), signal : Signal)
   }
   override def setLong(signal : Signal, value : Long) : Unit = {
-    assert(signal.id.toInt != -1, "You can't access this signal in the simulation, as it isn't public")
-    backend.nativeInstance.setU64(handle, signal.id.toInt, signal.dataType.longToRaw64(value, signal : Signal))
+    assert(signal.id != -1, "You can't access this signal in the simulation, as it isn't public")
+    backend.nativeInstance.setU64(handle, signal.id, signal.dataType.longToRaw64(value, signal : Signal))
   }
 
   override def getBigInt(signal: Signal) = {
     if(signal.dataType.width < 64 || (signal.dataType.width == 64 && signal.dataType.isInstanceOf[SIntDataType])) {
       getLong(signal)
     } else if(signal.dataType.width == 64){
-      val rawValue = backend.nativeInstance.getU64(handle, signal.id.toInt)
+      val rawValue = backend.nativeInstance.getU64(handle, signal.id)
       if(rawValue >= 0 ) {
         BigInt(rawValue)
       }else{
@@ -31,11 +31,11 @@ class SimVerilator(backend : VerilatorBackend, handle : Long) extends SimRaw(){
     } else {
       if(signal.dataType.isInstanceOf[SIntDataType]){
         val array = new Array[Byte]((signal.dataType.width+31)/32*4)
-        backend.nativeInstance.getAU8(handle, signal.id.toInt, array)
+        backend.nativeInstance.getAU8(handle, signal.id, array)
         BigInt(array)
       }else{
         val array = new Array[Byte]((signal.dataType.width+31)/32*4 + 1)
-        backend.nativeInstance.getAU8(handle, signal.id.toInt, array)
+        backend.nativeInstance.getAU8(handle, signal.id, array)
         array(0) = 0
         BigInt(array)
       }
@@ -47,14 +47,14 @@ class SimVerilator(backend : VerilatorBackend, handle : Long) extends SimRaw(){
     if(valueBitLength <= 63) {
       setLong(signal, value.toLong)
     } else if(valueBitLength == 64 && signal.dataType.width == 64) {
-      assert(signal.id.toInt != -1, "You can't access this signal in the simulation, as it isn't public")
+      assert(signal.id != -1, "You can't access this signal in the simulation, as it isn't public")
       val valueLong = value.toLong
       signal.dataType.checkIs64(valueLong, signal : Signal)
-      backend.nativeInstance.setU64(handle, signal.id.toInt, valueLong)
+      backend.nativeInstance.setU64(handle, signal.id, valueLong)
     } else {
       signal.dataType.checkBigIntRange(value, signal)
       val array = value.toByteArray
-      backend.nativeInstance.setAU8(handle, signal.id.toInt, array, array.length)
+      backend.nativeInstance.setAU8(handle, signal.id, array, array.length)
     }
   }
